Refactor: Remove unused widgets and implement new dialog and button components
- Deleted `message_dialog.dart` and `profile_app_bar.dart` as they were no longer needed.
- Introduced `custom_dialog.dart` and `custom_message_dialog.dart` for improved dialog handling.
- Added `custom_button.dart` with new button styles and functionalities.
- Created `custom_app_bar.dart` to consolidate app bar functionalities with rounded corners.
- Implemented `ReportButton` widget for reporting content with a streamlined dialog process.
